'Argument-less game start' bug fix
Typing "~4word" alone without an argument would try to set the word to
4word. This actually would succeed if a game was not yet in session
(somehow passing the spellcheck) and cause the game to start with
"4word". Not a good thing. The module currently will ignore when it
recieves its own name as a module - so avoid naming it a 4 letter
lowercase word.